+++ src/davotools/__init__.py
@@ -28,16 +28,8 @@ view.dict = _view.view_dict
 view.pd = _view.view_pd
 
 # %%
-## (internal) to set the module path
-def _set_path(echo: bool = True) -> tuple[str, str]:
-    module_path = os.path.abspath(__file__)
-    module_dir = os.path.dirname(module_path)
-    if echo:
-        print(f"-. davotools loaded from {module_dir}")
-    return module_path, module_dir
-
-# %%
 ## initialization
-module_path, module_dir = _set_path(echo=False)
+module_path = os.path.abspath(__file__)
+module_dir = os.path.dirname(module_path)
 
 # %%